+++ CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 **Note:** Patch versions prior to [v8.0.0](#800---2018-07-01) are not listed here.
 
 ## [Unreleased] - ????-??-??
+### Added
+- Event `start`: triggers when the message's payload is `{"command":"start"}` (when user pressed the `Start` button)
+- Accordingly, added the `start` event to stats
+
 ### Changed
 - Refactor statistics (use object to keep track of event counts)
+- Statistics no longer show empty at startup, instead `Stats initialized` shows up
 - Use `package.json` `"files"` field instead of `.npmignore`
 - Tarball will now include the source map